+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Configuration option `deactivateSmartyForCmsContent` to prevent Smarty from processing content added via CMS
+- Method `OxidEsales\EshopCommunity\Core\Model\BaseModel::getRawFieldData()`
 
 ### Changed
 - Update `symfony/expression-language` component
@@ -32,7 +33,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Text message on `Payment Methods > RDFa` tab
 - Docblock and other coding style fixes:
     - [PR-876](https://github.com/OXID-eSales/oxideshop_ce/pull/876)
-- Method `OxidEsales\EshopCommunity\Core\Model \BaseModel::getRawFieldData()` 
 
 ## [6.9.0] - 2021-07-27